+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVWWW4EPSTZZ8WR5YQSDB2AS"
+        "01KVX1B4JEBZWHAFGTNE2GC3PF"
     )
     r"""The unique identifier of the data source"""