+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVWWW4EHYE7WG0FYT233NN4M"
+        "01KVX1B4JAV5001JJZEATZZZ30"
     )
     r"""The unique identifier of the data source"""