+++ packages/orq-rc/RELEASES.md
@@ -8959,3 +8959,13 @@ Based on:
 - [python v4.11.0-rc.45] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.45] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.45 - packages/orq-rc
+
+## 2026-06-24 14:45:00
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.785.0 (2.912.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.46] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.46] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.46 - packages/orq-rc